클래스 나눠놓
